split package to scripts
