+++ test/accessor/dataframe/test_filter_in.py
@@ -1,0 +1,45 @@
+from test.accessor.conftest import d
+
+import pandas as pd
+import pytest
+
+import dtoolkit.accessor.dataframe  # noqa
+
+
+def test_work():
+    res = d.filter_in({"a": [0, 1], "b": [2]})
+
+    assert res["a"].isin([0, 1]).any()  # 0 and 1 in a
+    assert (~res["a"].isin([2])).all()  # 2 not in a
+    assert res["b"].isin([2]).any()  # 2 in a
+    assert (~res["b"].isin([0, 1])).all()  # 0 and not in a
+
+
+def test_inplace_is_true():
+    df = d.copy(True)
+    res = df.filter_in({"a": [0, 1], "b": [2]}, inplace=True)
+
+    assert res is None
+    assert df.equals(d) is False
+
+
+def test_issue_145():
+    # test my-data-toolkit#145
+    df = pd.DataFrame(
+        {
+            "legs": [2, 4, 2],
+            "wings": [2, 0, 0],
+        },
+        index=["falcon", "dog", "cat"],
+    )
+    res = df.filter_in({"legs": [2]})
+
+    expected = pd.DataFrame(
+        {
+            "legs": [2, 2],
+            "wings": [2, 0],
+        },
+        index=["falcon", "cat"],
+    )
+
+    assert res.equals(expected)
